Fix uninitialized memory error
See https://msdn.microsoft.com/en-us/library/windows/desktop/
 ms679351(v=vs.85).aspx

FORMAT_MESSAGE_ALLOCATE_BUFFER: The function allocates a buffer large
enough to hold the formatted message, and places a pointer to the
allocated buffer at the address specified by lpBuffer. The lpBuffer
parameter is a pointer to an LPTSTR; you must cast the pointer to an
LPTSTR (for example, (LPTSTR)&lpBuffer).